Revert "linux-karo: remove hand crafted defconfig and simply use that in karo meta layer"
This reverts commit 0cbabbc02008621bf1854fde3d5b8972df703c42.

Configuration in meta-karo lacks all of the virtualization settings and
therefore cannot be used as is.
This actoin was triggered by a wrong test [interpretation] (or strange
interaction of yocto build/sstate cache and replacing only the kernel on
the test device), which seemed to solve the performance issue.

Obviously I was fooled by Yocto which uses files/tx6/defconfig
regardless of being part of SRC_URI in bbappend file herein; probably
because of this very path which is already used in meta-karo layer and
Yocto merging all files together :-(
While the mechanism was not searched for in bitbake sources, a simple
test with the file being there or removed in conjunction with leaving it
out from SRC_URI definition showed proof of the behavior above.